Add Containerfile
Apparently I can name the Dockerfile as Containerfile and IMO it
makes more sense as Docker containers are not Docker specific.
Docker is only one of multiple container engines. A popular alt
one is Podman.
It only makes sense to not name this file after one of the multiple
available engines